Remove messagebox logic from utils and place its single usage   under debug/.
Simplify the messagebox logic to its most basic to facilitate
  further optimizations.
