--- src/coreclr/dlls/mscorrc/mscorrc.common.rc
@@ -1,9 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-STRINGTABLE DISCARDABLE
-BEGIN
-    IDS_EE_NAME_UNKNOWN                     "<Unknown>"
-    IDS_DEBUG_UNHANDLEDEXCEPTION             "Application has generated an exception that could not be handled.\n\nProcess ID=0x%x (%d), Thread ID=0x%x (%d).\n\nClick OK to terminate the application.\nClick CANCEL to debug the application."
-    IDS_DEBUG_SERVICE_CAPTION                "Application Error"
-END
