+++ src/coreclr/dlls/mscorrc/resource.h
@@ -167,9 +167,9 @@
 #define IDS_CLASSLOAD_STATICVIRTUAL_NOTIMPL     0x17a9
 
 #define IDS_CLASSLOAD_TOOMANYGENERICARGS        0x17ab
-#define IDS_DEBUG_SERVICE_CAPTION               0x17b4
+
 #define IDS_DEBUG_USERBREAKPOINT                0x17b6
-#define IDS_DEBUG_UNHANDLEDEXCEPTION            0x17b7
+
 #define IDS_PERFORMANCEMON_FUNCNOTFOUND         0x17bb
 #define IDS_PERFORMANCEMON_FUNCNOTFOUND_TITLE   0x17bc
 #define IDS_PERFORMANCEMON_PSAPINOTFOUND        0x17bd
@@ -223,7 +223,7 @@
 #define IDS_EE_STRUCTTOOCOMPLEX                 0x1a04
 #define IDS_EE_STRUCTARRAYTOOLARGE              0x1a05
 #define IDS_EE_BADMARSHALFIELD_NOSTRINGBUILDER  0x1a06
-#define IDS_EE_NAME_UNKNOWN                     0x1a07
+
 #define IDS_EE_NO_BACKING_CLASS_FACTORY         0x1a0b
 #define IDS_EE_STRING_TOOLONG                   0x1a0d
 #define IDS_EE_VARARG_NOT_SUPPORTED             0x1a0f
