+++ src/coreclr/vm/dwreport.cpp
@@ -953,17 +953,8 @@ void ResetWatsonBucketsFavorWorker(void * pParam)
     return;
 }
 
-// Forward declare the needed MessageBox API.
-int UtilMessageBoxCatastrophic(
-                  UINT uText,       // Text for MessageBox
-                  UINT uTitle,      // Title for MessageBox
-                  UINT uType,       // Style of MessageBox
-                  BOOL ShowFileNameInTitle, // Flag to show FileName in Caption
-                  ...);    // Additional Arguments
-
 //----------------------------------------------------------------------------
-// CreateThread() callback to invoke native Watson or put up our fake Watson
-// dialog depending on m_fDoReportFault value.
+// CreateThread() callback to invoke native Watson
 //
 // The output is a FaultReport* value communicated by setting
 // pFaultReportInfo->m_result. The DWORD function return value
@@ -984,32 +975,7 @@ static DWORD WINAPI DoFaultReportCreateThreadCallback(LPVOID pFaultReportInfoAsV
 
     FaultReportInfo *pFaultReportInfo = (FaultReportInfo*)pFaultReportInfoAsVoid;
     EXCEPTION_POINTERS *pExceptionInfo = pFaultReportInfo->m_pExceptionInfo;
-
-    if (pFaultReportInfo->m_fDoReportFault)
-    {
-        pFaultReportInfo->m_faultRepRetValResult = DoReportFault(pExceptionInfo);
-    }
-    else
-    {
-        int res = UtilMessageBoxCatastrophic(
-                               IDS_DEBUG_UNHANDLEDEXCEPTION,
-                               IDS_DEBUG_SERVICE_CAPTION,
-                               MB_OKCANCEL | MB_ICONEXCLAMATION,
-                               TRUE,
-                               GetCurrentProcessId(),
-                               GetCurrentProcessId(),
-                               pFaultReportInfo->m_threadid,
-                               pFaultReportInfo->m_threadid
-                              );
-        if (res == IDOK)
-        {
-            pFaultReportInfo->m_faultReportResult = FaultReportResultQuit;
-        }
-        else
-        {
-            pFaultReportInfo->m_faultReportResult = FaultReportResultDebug;
-        }
-    }
+    pFaultReportInfo->m_faultRepRetValResult = DoReportFault(pExceptionInfo);
 
     return 0;
 }
