+++ src/coreclr/vm/dwreport.h
@@ -84,10 +84,7 @@ extern LONG g_watsonAlreadyLaunched;
 typedef enum tagEFaultRepRetVal EFaultRepRetVal;
 struct FaultReportInfo
 {
-    BOOL                /*in*/   m_fDoReportFault;
     EXCEPTION_POINTERS  /*in*/  *m_pExceptionInfo;
-    DWORD               /*in*/   m_threadid;
-    FaultReportResult   /*out*/  m_faultReportResult;
     EFaultRepRetVal     /*out*/  m_faultRepRetValResult;
 };
